refactor(lists): :recycle: Separate lists api routes from client routes
Refs #12
