+++ CHANGELOG.md
@@ -15,8 +15,8 @@ File has been auto generated by version-bump. See
 ### Features
 
 - :lipstick: Custom CSS file
-  [831f751b](https://github.com/ultraxlight/app/commit/831f751b0c976d88e125a6d00260f907caacb65d)## [0.8.1](https://github.com/ultraxlight/app/compare/0.8.0..0.8.1)
-
+  [831f751b](https://github.com/ultraxlight/app/commit/831f751b0c976d88e125a6d00260f907caacb65d)##
+  [0.8.1](https://github.com/ultraxlight/app/compare/0.8.0..0.8.1)
 
 ### Other
 
@@ -24,7 +24,6 @@ File has been auto generated by version-bump. See
   [07d007dc](https://github.com/ultraxlight/app/commit/07d007dc627a61f46355119260989341dee5494e)
 
 ## [0.8.0](https://github.com/ultraxlight/app/compare/0.7.0..0.8.0)
-
 
 ### Features
